MyProject/src/main/java/com/mystore/pageobjects/SearchResultPage.java
/**
 * 
 */
package com.mystore.pageobjects;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.mystore.actiondriver.Action;
import com.mystore.base.BaseClass;

/**
 * @author Priyanka M
 *
 */
public class SearchResultPage extends BaseClass {
	
	@FindBy(xpath="//img[@title='Faded Short Sleeve T-shirts']") 
	WebElement productResult;
	
	public SearchResultPage() {
		PageFactory.initElements(getDriver(), this);
		
	}
	
	public boolean isproductAvailable() {
		return Action.isDisplayed(getDriver(), productResult);
		
	}
	
	public AddToCartPage clickOnProduct() {
			Action.click(getDriver(), productResult);
			return new AddToCartPage();
	}

}
